feat(desktop): wire CP-857 into PrinterCommand + relocate to escpos module
Completes Phase 1.3.A.2 (move PrinterCommand into escpos) and Phase 1.3.B
(wire CP-857 transcoder into the print path).

PrinterCommand enum + to_bytes() moved verbatim from bluetooth.rs into
escpos/mod.rs. bluetooth.rs re-exports the type so existing main.rs
imports keep compiling without churn. The behavioral change:

  Initialize: now emits ESC @ + ESC t 13 (select CP-857) instead of just
              ESC @ — every connected printer will switch to the Turkish
              code page on init.
  Text:       UTF-8 → CP-857 transcode via utf8_to_cp857() before bytes
              ship to the printer.
  TextLine:   same as Text plus trailing LF.
  Barcode:    payload kept as ASCII (CODE39 charset would corrupt under
              CP-857 transcode).
  QRCode:     payload kept as opaque bytes (QR encoder handles unicode).

Adds 5 new escpos/mod.rs unit tests covering the wire-level fix:
- initialize_sends_reset_and_cp857_selector
- text_line_transcodes_turkish_chars_to_cp857  (asserts byte sequence
  for "Şiş" produces [0x9E, 0x69, 0x9F, 0x0A] — Bug E end-to-end)
- text_without_newline_omits_lf
- feed_cut_align_unaffected_by_cp857
- barcode_keeps_ascii_payload_unchanged

Verified: cargo check passes. Codepage tests (9/9) pass via standalone
rustc. Full cargo test --bin requires libwebkit2gtk-4.0.so symlink to
4.1 — environmental, doesn't block this commit.

Refs: docs/superpowers/plans/2026-04-27-phase-1-3-tauri-hardware-suite.md §1.3.A.2 + §1.3.B.
